Break PreferOutgoingHandler out of peer module
